requirements.txt
asgiref==3.8.1
bleach==4.1.0
cloudinary==1.40.0
crispy-bootstrap5==2024.2
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==4.2.13
django-allauth==0.57.2
django-crispy-forms==2.1
django-jquery==3.1.0
django-localflavor==4.0
django-summernote==0.8.20.0
django-tables2==2.7.0
gunicorn==20.1.0
oauthlib==3.2.2
psycopg2==2.9.9
PyJWT==2.8.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-stdnum==1.20
python3-openid==3.2.0
requests-oauthlib==2.0.0
sqlparse==0.5.0
urllib3==1.26.18
whitenoise==5.3.0
